Move all forms to package org.kitodo.forms
