+++ Kitodo/src/main/java/de/sub/goobi/helper/tasks/CreateNewspaperProcessesTask.java
@@ -11,7 +11,6 @@
 
 package de.sub.goobi.helper.tasks;
 
-import de.sub.goobi.forms.ProzesskopieForm;
 import de.sub.goobi.helper.Helper;
 
 import java.io.IOException;
@@ -43,6 +42,7 @@ import org.kitodo.data.database.beans.Batch.Type;
 import org.kitodo.data.database.beans.Process;
 import org.kitodo.data.exceptions.DataException;
 import org.kitodo.exceptions.ProcessCreationException;
+import org.kitodo.forms.ProzesskopieForm;
 import org.kitodo.services.ServiceManager;
 
 /**
