+++ Kitodo/src/main/java/de/sub/goobi/helper/tasks/TaskManager.java
@@ -30,7 +30,7 @@ import org.kitodo.config.Parameters;
 /**
  * The class TaskManager serves to handle the execution of threads. It can be
  * user controlled by the “Long running task manager”, backed by
- * {@link de.sub.goobi.forms.LongRunningTasksForm}.
+ * {@link org.kitodo.forms.LongRunningTasksForm}.
  *
  * @author Matthias Ronge &lt;matthias.ronge@zeutschel.de&gt;
  */
