+++ Kitodo/src/main/java/org/goobi/mq/processors/CreateNewProcessProcessor.java
@@ -12,8 +12,6 @@
 package org.goobi.mq.processors;
 
 import de.sub.goobi.config.ConfigCore;
-import de.sub.goobi.forms.AdditionalField;
-import de.sub.goobi.forms.ProzesskopieForm;
 import de.sub.goobi.helper.Helper;
 import de.unigoettingen.sub.search.opac.ConfigOpacDoctype;
 
@@ -38,6 +36,8 @@ import org.kitodo.api.ugh.exceptions.WriteException;
 import org.kitodo.config.Parameters;
 import org.kitodo.data.database.beans.Template;
 import org.kitodo.exceptions.ProcessCreationException;
+import org.kitodo.forms.ProzesskopieForm;
+import org.kitodo.helper.AdditionalField;
 import org.kitodo.services.ServiceManager;
 
 /**
