+++ Kitodo/src/main/java/org/kitodo/forms/BaseForm.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package de.sub.goobi.forms;
+package org.kitodo.forms;
 
 import java.io.Serializable;
 import java.util.List;
