+++ Kitodo/src/main/java/org/kitodo/forms/LdapGruppenForm.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package de.sub.goobi.forms;
+package org.kitodo.forms;
 
 import de.sub.goobi.helper.Helper;
