+++ Kitodo/src/main/java/org/kitodo/forms/LdapServerForm.java
@@ -11,7 +11,6 @@
 
 package org.kitodo.forms;
 
-import de.sub.goobi.forms.BaseForm;
 import de.sub.goobi.helper.Helper;
 
 import java.text.MessageFormat;
