+++ Kitodo/src/main/java/org/kitodo/forms/ProzessverwaltungForm.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package de.sub.goobi.forms;
+package org.kitodo.forms;
 
 import com.lowagie.text.Document;
 import com.lowagie.text.DocumentException;
