+++ Kitodo/src/main/java/org/kitodo/forms/WorkflowForm.java
@@ -12,7 +12,6 @@
 package org.kitodo.forms;
 
 import de.sub.goobi.config.ConfigCore;
-import de.sub.goobi.forms.BaseForm;
 import de.sub.goobi.helper.Helper;
 
 import java.io.BufferedReader;
