+++ Kitodo/src/test/java/org/kitodo/forms/IndexingFormIT.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package de.sub.goobi.forms;
+package org.kitodo.forms;
 
 import org.junit.AfterClass;
 import org.junit.Assert;
