+++ Kitodo/src/test/java/org/kitodo/forms/ProzesskopieFormTest.java
@@ -9,7 +9,7 @@
  * GPL3-License.txt file that was distributed with this source code.
  */
 
-package de.sub.goobi.forms;
+package org.kitodo.forms;
 
 import static org.junit.Assert.assertEquals;
